requirements.txt
# certifi
aiohttp
json-stream
pysimdjson
tomlkit
libusb1
cryptography==41.0.0
colorama
readchar
tableprint
typing_extensions
